# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ReceiptBud API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /receipts:
    get:
      operationId: listReceipts
      tags: [receipts]
      summary: List receipts
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: favorite
          in: query
          schema:
            type: boolean
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
        - name: sortOrder
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of receipts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Receipt"
    post:
      operationId: createReceipt
      tags: [receipts]
      summary: Create a receipt manually or store scanned data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReceiptInput"
      responses:
        "201":
          description: Created receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Receipt"
  /receipts/scan:
    post:
      operationId: scanReceipt
      tags: [receipts]
      summary: Extract receipt data from an image using AI (base64)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReceiptScanInput"
      responses:
        "200":
          description: Extracted receipt data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReceiptScanResult"
  /receipts/{id}:
    get:
      operationId: getReceipt
      tags: [receipts]
      summary: Get a receipt by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Receipt"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateReceipt
      tags: [receipts]
      summary: Update a receipt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReceiptUpdate"
      responses:
        "200":
          description: Updated receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Receipt"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteReceipt
      tags: [receipts]
      summary: Delete a receipt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /receipts/{id}/favorite:
    patch:
      operationId: toggleReceiptFavorite
      tags: [receipts]
      summary: Toggle receipt favorite status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Receipt"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ── Budgets ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── Receipt schemas ──────────────────────────────────────────────────────
    ReceiptItem:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
        quantity:
          type: ["number", "null"]
        category:
          type: ["string", "null"]
      required:
        - name
        - price
    Receipt:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        storeName:
          type: string
        date:
          type: string
        total:
          type: number
        tax:
          type: ["number", "null"]
        discount:
          type: ["number", "null"]
        paymentMethod:
          type: ["string", "null"]
        category:
          type: string
        notes:
          type: ["string", "null"]
        imageBase64:
          type: ["string", "null"]
        isFavorite:
          type: boolean
        items:
          type: array
          items:
            $ref: "#/components/schemas/ReceiptItem"
        currency:
          type: string
          default: "USD"
        aiInsight:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - storeName
        - date
        - total
        - category
        - currency
        - isFavorite
        - items
        - createdAt
        - updatedAt
    ReceiptInput:
      type: object
      properties:
        storeName:
          type: string
        date:
          type: string
        total:
          type: number
        tax:
          type: number
        discount:
          type: number
        paymentMethod:
          type: string
        category:
          type: string
        notes:
          type: string
        imageBase64:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/ReceiptItem"
        currency:
          type: string
        aiInsight:
          type: string
      required:
        - storeName
        - date
        - total
        - category
        - items
    ReceiptUpdate:
      type: object
      properties:
        storeName:
          type: string
        date:
          type: string
        total:
          type: number
        tax:
          type: number
        discount:
          type: number
        paymentMethod:
          type: string
        category:
          type: string
        notes:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/ReceiptItem"
    ReceiptScanInput:
      type: object
      properties:
        imageBase64:
          type: string
        mimeType:
          type: string
      required:
        - imageBase64
    ReceiptScanResult:
      type: object
      properties:
        storeName:
          type: string
        date:
          type: string
        total:
          type: number
        tax:
          type: ["number", "null"]
        discount:
          type: ["number", "null"]
        paymentMethod:
          type: ["string", "null"]
        category:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/ReceiptItem"
        aiInsight:
          type: ["string", "null"]
      required:
        - storeName
        - date
        - total
        - category
        - items

    # ── Budget schemas ───────────────────────────────────────────────────────
...